backend/requirements.txt
fastapi
uvicorn

#for databases
sqlalchemy
psycopg2

#for env variables
python-dotenv